feat: get commits between last release and HEAD
Compares the last release commit with HEAD (tip of default branch) to
get a status (ahead, behind, diverged, or identical), and the messages
from the commits that separate them. These messages along with the
version extracted from the release will be used to calculate the version
and changelog for the next release.