scratch: show Insomniak47 the VersionTagParser
I lost the plot a bit when I made the --base option parse a Version
instead of just confirming that a version COULD be extracted from it;
but the code is interesting so I'm sharing it anyway. What should
actually happen is we should confirm that we're able to extract a
version but keep the passed argument as a string because we need to use
it in the URI of the request for the release. We only want to pull the
version out to use as the base when calculating the new version for the
version subcommand.